source/requirements.txt
psycopg2-binary==2.9.5
django-bootstrap5==22.2
Django==4.1.7
pillow==10.3.0
django-ckeditor-5==0.2.13
gunicorn
requests==2.32.3

djangorestframework==3.14
drf_spectacular==0.24.1
drf-spectacular-sidecar==2022.12.1

